Large change to display code to avoid using xbl
Using methods on xbl widgets is unreliable and not worth the hassle. It's much easier to keep all the code in one place, ie: pop.js

This patch removes the dependency on the xbl widget for the results (but doesn't actually remove the unused xbl code yet).

It's a large change and introduces a whole load of unfixed bugs.