Add allow_external_via_proxy option
Requires user to enable use of external tools when a proxy is set. This is
disabled by default because it can harm users of anonymising proxies: there's
likely to be a "what's your real ip address" question the remote end can send
for at least one of rtmp/mms/whatever.